TEMPORARY: added a time tracking in solver.py within transient_solve_TR() function which tracks the time of solver()
This is not necessary at all, just for me to remember how I did it before. I'm going to track the time of the solver whenever it is called.
Note: This does not make a lot of sense right now yet, because the solver is called to solve the heat diffusion equation as well as firn gas diffusion and also water isotope diffusion (this is discarded in my experiments). I still need to separate this from each other.