Including numba. Does not work yet!!! The code runs through, but is slower than without jit.
Numba can not deal with dictionaries, so I rewrote w() such that it does not get the dictionary "airdict" but all the necessary variables from the dictionary.
I also calculate gradients before w() is called because numba does not support np.gradient();
furthermore w() calls A() which uses np.size, which is not supported. Use a.size instead which seems to work (no clue why)